% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_lock_exam.R
\name{exam_check_ui}
\alias{exam_check_ui}
\title{Check Exam Submission}
\usage{
exam_check_ui(id, label = "Check submissions")
}
\arguments{
\item{id}{ID matching ui with "lock_server()"}

\item{label}{Label for view button}
}
\description{
View if any questions or exercises have not been submitted prior to locking the exam.
}
