% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_option.R
\name{exam_output_ui}
\alias{exam_output_ui}
\title{View exam submission output}
\usage{
exam_output_ui(id, label = "View submissions")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label for view button}
}
\description{
Obtain output for exam.
}
