% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_grade_tutorial.R
\name{grade_button_ui}
\alias{grade_button_ui}
\title{Tutorial grade button}
\usage{
grade_button_ui(id, label = "View Grade")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the submit grade button}
}
\description{
View current grade and submission attempt on all question and exercises for a user.
This is used so users can check their progress when working on a tutorial
that allows for multiple attempts, immediate feedback, and deductions after x attempts.
The user can download their grade to an html by including the \code{grade_print_ui()} function.

Shiny ui and server logic for the grade computation.

Note that when including these functions in a learnr Rmd document it is necessary that
the server function, \code{grade_server()}, be included in an R chunk where \code{context="server"} as
they interact with the underlying Shiny functionality. Conversely, the ui function,
\code{grade_button_ui()}, must \emph{not} be included in an R chunk with a \code{context}.
}
