% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_exam_server}
\alias{grade_exam_server}
\title{Tutorial grade server}
\usage{
grade_exam_server(id, label = NULL, pts_possible = NULL, exclude = NULL)
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Either NULL or a vector containing the names of each question/exercise.}

\item{pts_possible}{Either NULL or a vector containing the number of points corresponding to each question/exercise in label.}

\item{exclude}{Either NULL or a vector of names of questions/exercises to exclude.}
}
\description{
Tutorial grade server
}
