% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_exam_ui}
\alias{grade_exam_ui}
\title{Exam grade button}
\usage{
grade_exam_ui(id, label = "Download Exam")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the submit grade button}
}
\description{
print an exam
}
