% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_output_ui}
\alias{grade_output_ui}
\title{Tutorial grade output}
\usage{
grade_output_ui(id)
}
\arguments{
\item{id}{ID matching ui with server}
}
\description{
Obtain grade on all question and exercise submissions for a user.
}
