% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_print_ui}
\alias{grade_print_ui}
\title{Tutorial print grade}
\usage{
grade_print_ui(id, label = "Download Grade")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the button}
}
\description{
Obtain grade on all question and exercise submissions for a user.
}
