% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_server}
\alias{grade_server}
\title{Tutorial grade server}
\usage{
grade_server(id, rubric = NULL, num_try = 3, deduction = 0.1, name = TRUE)
}
\arguments{
\item{id}{ID matching ui with server}

\item{rubric}{Either NULL or a data frame containing a vector of question names and a vector of pts_possible for each question}

\item{num_try}{Number of tries allowed before grade deduction on that question. Default is 3.}

\item{deduction}{The percent (as a decimal) to be deducted for each additional incorrect attempt after num_try. Default is 0.1.}

\item{name}{Have "Name" count towards grade? TRUE or FALSE}
}
\description{
Tutorial grade server
}
