% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_quiz.R
\name{knit_print.tutorial_question}
\alias{knit_print.tutorial_question}
\title{Knitr quiz print methods}
\usage{
\method{knit_print}{tutorial_question}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
Print methods for question and exam functions
}
\keyword{internal}
