% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_lock_exam.R
\name{lock_check_ui}
\alias{lock_check_ui}
\title{Check Exam Submission}
\usage{
lock_check_ui(id, label = "Check Submissions")
}
\arguments{
\item{id}{ID matching ui with "lock_server()"}

\item{label}{Label for button}
}
\description{
View if any questions or exercises have not been submitted prior to locking the exam.
}
