% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_dropdown.R
\name{question_dropdown}
\alias{question_dropdown}
\title{Drop-down question for learnr tutorials.}
\usage{
question_dropdown(
  text,
  ...,
  type = "dropdown",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = sortable::sortable_options()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{type}{Type of quiz question. Typically this can be automatically
determined based on the provided answers. Pass \code{"radio"} to indicate that
even though multiple correct answers are specified that inputs which
include only one correct answer are still correct. Pass \code{"checkbox"} to
force the use of checkboxes (as opposed to radio buttons) even though only
one correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to
"Incorrect") when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer when \code{allow_retry}
is \code{TRUE}.
Defaults to "Incorrect. Be sure to select every correct answer." for
checkbox questions and "Incorrect" for non-checkbox questions.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Extra options to be stored in the question object. This is
useful when using custom question types. See \code{\link[sortable:question_rank]{sortable::question_rank()}}
for an example question implementation that uses the \code{options} parameter.}
}
\value{
A custom \code{learnr} question, with \code{type = dropdown}.
}
\description{
Add interactive drop-down tasks to your \code{learnr} tutorials.
An alternative option to \code{question_radio} when you have a lot of options
and want to save space.
}
\examples{
question_dropdown(
  "Pick the letter B",
  learnr::answer("A"),
  learnr::answer("B", correct = TRUE),
  learnr::answer("C"),
  learnr::answer("D"),
  allow_retry = TRUE,
  random_answer_order = TRUE
)

}
