% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_option.R
\name{reset_server}
\alias{reset_server}
\title{Tutorial reset server}
\usage{
reset_server(id, file_name = NULL, package_name = NULL)
}
\arguments{
\item{id}{ID matching ui with server}

\item{file_name}{Name of the .Rmd file (not the tutorial id)}

\item{package}{Name of the package with tutorial.}
}
\description{
Tutorial reset server
}
