% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{text_server}
\alias{text_server}
\title{Text and description server}
\usage{
text_server(id, text = NULL)
}
\arguments{
\item{id}{ID matching ui with server}

\item{text}{output to save and print}
}
\description{
Text and description server
}
