% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_button_ui}
\alias{grade_button_ui}
\title{Tutorial grade button}
\usage{
grade_button_ui(id, label = "grade tutorial")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the submit grade button}
}
\description{
Obtain grade on all question and exercise submissions for a user.

Shiny ui and server logic for the grade computation.

Note that when including these functions in a learnr Rmd document it is necessary that
the server function, \code{grade_server()}, be included in an R chunk where \code{context="server"} as
they interact with the underlying Shiny functionality. Conversely, the ui function,
\code{grade_ui()}, must \emph{not} be included in an R chunk with a \code{context}.
}
