% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_grade_tutorial.R
\name{grade_print_ui}
\alias{grade_print_ui}
\title{Tutorial print grade}
\usage{
grade_print_ui(id, label = "Download Grade")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the button}
}
\description{
Download an html grade of the user's tutorial.

Shiny ui and server logic for the grade computation.

Note that when including these functions in a learnr Rmd document it is necessary that
the server function, \code{grade_server()}, be included in an R chunk where \code{context="server"} as
they interact with the underlying Shiny functionality. Conversely, the ui function,
\code{grade_print_ui()}, must \emph{not} be included in an R chunk with a \code{context}.
}
