% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_server}
\alias{grade_server}
\title{Tutorial grade server}
\usage{
grade_server(
  id,
  label = NULL,
  pts_possible = NULL,
  num_try = 3,
  deduction = 0.1,
  exclude = NULL
)
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Either NULL or a vector containing the names of each question/exercise.}

\item{pts_possible}{Either NULL or a vector containing the number of points corresponding to each question/exercise in label.}

\item{num_try}{Number of tries allowed before grade deduction on that question. Default is 3.}

\item{deduction}{The percent (as a decimal) to be deducted for each additional incorrect attempt after num_try. Default is 0.1.}

\item{exclude}{Either NULL or a vector of names of questions/exercises to exclude.}
}
\description{
Tutorial grade server
}
