% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_option.R
\name{lock_button_ui}
\alias{lock_button_ui}
\title{Tutorial lock button}
\usage{
lock_button_ui(id, label = "lock exam")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the submit grade button}
}
\description{
Lock all "submit" buttons.

Shiny ui and server logic for the lock computation.

Note that when including these functions in a learnr Rmd document it is necessary that
the server function, \code{lock_server()}, be included in an R chunk where \code{context="server"} as
they interact with the underlying Shiny functionality. Conversely, the ui function,
\code{lock_button_ui()}, must \emph{not} be included in an R chunk with a \code{context}.
}
