% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_option.R
\name{lock_server}
\alias{lock_server}
\title{Exam lock and grade server}
\usage{
lock_server(
  id,
  num_blanks = TRUE,
  show_correct = FALSE,
  graded = NULL,
  graded_pts = NULL,
  ex = NULL,
  ex_pts = NULL,
  manual = NULL,
  manual_pts = NULL,
  exclude = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{id}{ID matching ui with server}

\item{num_blanks}{TRUE/FALSE: Set the number of points for a question equal to the number of blanks. Default TRUE.}

\item{show_correct}{TRUE/FALSE: Whether or not to show points for each question in grade output. Default TRUE.}

\item{graded}{Either NULL or a vector containing the names of each question/exercise.}

\item{graded_pts}{Either NULL or a vector containing the number of points corresponding to each question/exercise in graded.}

\item{ex}{Either NULL or a vector containing the names of each question/exercise.}

\item{ex_pts}{Either NULL or a vector containing the number of points corresponding to each question/exercise in ex.}

\item{manual}{Either NULL or a vector containing the names of each question/exercise to be manually graded.}

\item{manual_pts}{Either NULL or a vector containing the number of points corresponding to each question/exercise in manual.}

\item{exclude}{Either NULL or a vector containing the names of each question/exercise to exclude from grading.}

\item{tz}{Time zone to display start time on report.}
}
\description{
Exam lock and grade server
}
