% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{Tutorial exam}
\usage{
exam(..., caption = rlang::missing_arg(), shuffle = FALSE)
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}

\item{shuffle}{Randomly shuffle the questions within quiz}
}
\description{
Add interactive exam questions to a tutorial.
}
