% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_retry.R
\name{reset_server}
\alias{reset_server}
\title{Tutorial reset server}
\usage{
reset_server(id, file_name = NULL, package_name = NULL, tz = Sys.timezone())
}
\arguments{
\item{id}{ID matching ui with server}

\item{file_name}{Tutorial name (subdirectory within tutorials/directory of installed package)}

\item{package_name}{Name of package with tutorial.}

\item{tz}{Time zone to display retry time.}
}
\description{
Tutorial reset server
}
